i45: Combine related actions and reducers
