+++ index.js
@@ -7,7 +7,8 @@ import { AppRegistry, AppState } from 'react-native';
 import auth from '@react-native-firebase/auth';
 import store from 'rdx/store';
 import { Provider } from 'react-redux';
-import { signIn, updateAuth, updateAppState, getStateFromLS } from 'rdx/actions';
+import { updateAppState, getStateFromLS } from 'rdx/appState';
+import { signIn, updateAuth } from 'rdx/authState';
 import { SafeAreaProvider, initialWindowMetrics } from 'react-native-safe-area-context';
 
 import App from './App';
@@ -16,13 +17,8 @@ import { name as appName } from './app.json';
 const AppContainer = () => {
   useEffect(() => {
     store.dispatch(getStateFromLS());
-    const unsubscribe = auth().onAuthStateChanged((user) => {
-      if (user) {
-        store.dispatch(signIn(user));
-      } else {
-        store.dispatch(updateAuth(false));
-      }
-    });
+    const unsubscribe = auth().onAuthStateChanged((user) =>
+      (user ? store.dispatch(signIn(user)) : store.dispatch(updateAuth(false))));
 
     AppState.addEventListener('change', appStateHandler);
