+++ src/library/components/HomeBar.js
@@ -2,7 +2,8 @@ import React, { useState } from 'react';
 import { View, Image, TouchableOpacity } from 'react-native';
 import { connect } from 'react-redux';
 import auth from '@react-native-firebase/auth';
-import { signOut, goToUser } from 'rdx/actions';
+import { goToUser } from 'rdx/locationState';
+import { signOut } from 'rdx/authState';
 import { getLocation } from 'library/apis/geolocation';
 import { create } from 'library/utils/normalize.js';
 import { Menu, MenuDivider, MenuItem } from './Menu';
