--- src/redux/reducers/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import authReducer from './authReducer';
-import mapReducer from './mapReducer';
-import appReducer from './appReducer';
-
-export default combineReducers({
-  auth: authReducer,
-  location: mapReducer,
-  app: appReducer,
-});
