+++ src/redux/store.js
@@ -1,9 +1,9 @@
 import { createStore, applyMiddleware } from 'redux';
 import reduxThunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import stateValidator from 'rdx/middlewares/stateValidator';
-import reducers from 'rdx/reducers';
-import INITIAL_STATE from 'rdx/initialState';
+import stateValidator from './middlewares/stateValidator';
+import reducers from './reducers';
+import INITIAL_STATE from './initialState';
 
 const composeEnhancers = composeWithDevTools({
   trace: true,
